compiler version info + locked pragmas
Contracts should always use locked compiler version pragmas without
a caret prefix. The near-match caret `^` only checks that major and minor versions match, in our case 0.4.x. If a dev has the newest solidity version, 0.4.16, installed on their machine, the code with near-match pragmas will compile, but with different bytecode and thus different behavior than other dev who uses `solc` 0.4.15 locally. This scenario is impossible when the carets are removed. We must be sure all testing and deployment is done with the same version of `solc`.